[Frontend] Move LoRA endpoints out of openai api_server and SageMaker specific endpoints out of openai api_server.
